==> titanic/notebooks/1. Extracting_Titanic_Data.ipynb <==
!pip install python-dotenv

from dotenv import load_dotenv,find_dotenv

# find .env file automatically by walking up the directories
dotenv_path = find_dotenv()
#load up the entries as environment variables
load_dotenv(dotenv_path)

import os
KAGGLE_USERNAME = os.environ.get('KAGGLE_USERNAME')
print KAGGLE_USERNAME

import requests
from requests import session
import os
from dotenv import load_dotenv,find_dotenv

#payload for post
payload = {
    'action': 'login',
    'username': os.environ.get('KAGGLE_USERNAME'),
    'password': os.environ.get('KAGGLE_PASSWORD')    
}

# Url for train data of Titanic from kaggle
url = 'https://www.kaggle.com/c/3136/download/train.csv/'

#setup session
with session() as c:
    #post request
    c.post('https://kaggle.com/account/login', data=payload)
    #get request
    response = c.get(url)
    
    print(response.text)
