==> titanic/notebooks/3.Building_Predictive_Models.ipynb <==
import pandas as pd
import numpy as np
import os

### Import Data

#Set file path for train and test data
file_path = "C:/Users/ABHISHEK/titanic/data/processed/"
train_file_name = "train.csv"
test_file_name = "test.csv"
train_path = os.path.join(file_path,train_file_name)
test_path = os.path.join(file_path,test_file_name)

#read the data with all default parameters
#PassengerId is the unique parameter among all columns - Primary Key to be clear
train_df = pd.read_csv(train_path,index_col = "PassengerId")
test_df = pd.read_csv(test_path,index_col="PassengerId")

train_df.info()

train_df.columns

test_df.info()

### Data Preparation

X = train_df.loc[:,'Age':].as_matrix().astype('float')
y = train_df['Survived'].ravel()

print X.shape, y.shape

# train test split
from sklearn.model_selection import train_test_split
X_train,X_test,y_train,y_test = train_test_split(X, y, test_size=0.2,random_state=0)
print X_train.shape, y_train.shape
print X_test.shape, y_test.shape

# average survival in train and test
print "mean survival in train : {0:.3f}".format(np.mean(y_train))
print "mean survival in test : {0:.3f}".format(np.mean(y_test))
# This is necessary to check if the positive values are evenly distributed in train and test data
# Only 39% approx positive values are appearing. If it is even less or much high, there is a different approach for such
# unbalanced data. Now we go ahead and build the baseline model.

import sklearn

sklearn.__version__

### BaseLine Model

from sklearn.dummy import DummyClassifier

# create model
model_dummy = DummyClassifier(strategy='most_frequent', random_state=0)

# train model
model_dummy.fit(X_train, y_train)

print('score for baseline model : {0:.2f}'.format(model_dummy.score(X_test, y_test)))

# performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

# accuracy score
print('accuracy for baseline model : {0:.2f}'.format(accuracy_score(y_test,model_dummy.predict(X_test))))

# confusion matrix
print('confusion matrix for baseline model : \n {0}'.format(confusion_matrix(y_test,model_dummy.predict(X_test))))

# precision and recall scores
print('precision score for baseline model : {0:.2f}'.format(precision_score(y_test,model_dummy.predict(X_test))))
print('recall score for baseline model : {0:.2f}'.format(recall_score(y_test,model_dummy.predict(X_test))))

### First Kaggle Submission

# converting to matrix
test_X = test_df.as_matrix().astype('float')

# get predictions
predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})

df_submission.head()

submission_data_path = "C:/Users/ABHISHEK/titanic/data/external"
submission_file_path = os.path.join(submission_data_path,'01_dummy.csv')

df_submission.to_csv(submission_file_path,index=False)

def get_submission_file(model,filename):
    # converting to matrix
    test_X = test_df.as_matrix().astype('float')
    # make predictions
    predictions = model.predict(test_X)
    # submission dataframe
    df_submission = pd.DataFrame({'PassengerId': test_df.index, 'Survived': predictions})
    # submission file path
    submission_data_path = "C:/Users/ABHISHEK/titanic/data/external"
    submission_file_path = os.path.join(submission_data_path,filename)
    # write to file
    df_submission.to_csv(submission_file_path,index=False)

get_submission_file(model_dummy,'01_dummy.csv')

### Logistic Regression Model

from sklearn.linear_model import LogisticRegression

# create model
model_lr_1 = LogisticRegression(random_state=0)

# train model
model_lr_1.fit(X_train,y_train)

# Accuracy for logistic regression
print('score for logistic regession model - version - 1 : {0:.2f}'.format(model_lr_1.score(X_test, y_test)))

# Performance metrics for Logistic Regression
# accuracy score
print('accuracy for logistic regression model 1: {0:.2f}'.format(accuracy_score(y_test,model_lr_1.predict(X_test))))
# confusion matrix
print('confusion matrix for logistic regression model 1 : \n {0}'.format(confusion_matrix(y_test,model_lr_1.predict(X_test))))
# precision and recall scores
print('precision score for logistic regression model 1 : {0:.2f}'.format(precision_score(y_test,model_lr_1.predict(X_test))))
print('recall score for logistic regression model 1 : {0:.2f}'.format(recall_score(y_test,model_lr_1.predict(X_test))))

# model coefficients
model_lr_1.coef_

### Second Kaggle Submission

# get submission file
get_submission_file(model_lr_1, '02_lr.csv') 

### Hyperparamater Optimization

# base model
model_lr = LogisticRegression(random_state=0)

from sklearn.model_selection import GridSearchCV

parameters = {'C':[1.0,10.0,50.0,100.0,1000.0],'penalty':['l1','l2']}
clf = GridSearchCV(model_lr,param_grid=parameters,cv=3)
# Performs grid search for all combinations of given parameters and 3-fold cross validation too(since cv=3)

clf.fit(X_train, y_train)

clf.best_params_

print "best score : {0:.2f}".format(clf.best_score_)

# evaluate model
print('score for logistic regession model - version - 2 : {0:.2f}'.format(clf.score(X_test, y_test)))

### Third Kaggle Submission

get_submission_file(clf,'03_lr.csv')

### Feature normalization and standardization

from sklearn.preprocessing import StandardScaler, MinMaxScaler

#### Feature Normalization

scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_test_scaled = scaler.transform(X_test)

#### Feature Standardization

scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_test_scaled = scaler.transform(X_test)
X_scaled = scaler.fit_transform(X)

### Create Model after Standardization

# base model
model_lr = LogisticRegression(random_state=0)
parameters = {'C':[1.0,10.0,50.0,100.0,1000.0],'penalty':['l1','l2']}
clf = GridSearchCV(model_lr,param_grid=parameters,cv=3)
# Performs grid search for all combinations of given parameters and 3-fold cross validation too(since cv=3)

clf.fit(X_train_scaled, y_train)

clf.best_params_

print "best score : {0:.2f}".format(clf.best_score_)

# evaluate model
print('score for logistic regession model - version - 2 : {0:.2f}'.format(clf.score(X_test_scaled, y_test)))

### Model Persistence

# Helps in saving and loading models to make predictions. Also can be built into an API
import pickle

# create a file path
model_file_path = "C:/Users/ABHISHEK/titanic/models/lr_model.pkl"
scaler_file_path = "C:/Users/ABHISHEK/titanic/models/lr_scaler.pkl"

# open files to write
model_file_pickle = open(model_file_path,'wb')
scaler_file_pickle = open(scaler_file_path,'wb')

# persist the model and scaler
pickle.dump(clf,model_file_pickle)
pickle.dump(scaler,scaler_file_pickle)

# close the file
model_file_pickle.close()
scaler_file_pickle.close()

#### load the persisted file

# open files in read mode
model_file_pickle = open(model_file_path,'r')
scaler_file_pickle = open(scaler_file_path,'r')
# load files
clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)

clf_loaded

scaler_loaded

# transform the test data using loaded scaler object
X_test_scaled = scaler_loaded.transform(X_test)
# calculate score using loaded model object
print 'score for persisted logistic regression : {0:.2f}'.format(clf_loaded.score(X_test_scaled,y_test))

final_lr = LogisticRegression(C=1.0,penalty='l1',random_state=0)
final_lr.fit(X_scaled,y)
get_submission_file(final_lr,'03_lr_final.csv')
