==> titanic/notebooks/4.Building_ML_API.ipynb <==
hello_world_script_file = "C:/Users/ABHISHEK/titanic/src/models/hello_world_api.py"

%%writefile $hello_world_script_file

from flask import Flask, request

app = Flask(__name__)

@app.route('/api', methods=['POST'])
def say_hello():
    data = request.get_json(force=True)
    name = data['name']
    return 'hello {0}'.format(name)

if__name__ == '__main__':
    app.run(port=10001,debug=True)

import json
import requests

url = 'http://192.168.1.16:10001/api' # give local host ip address
data = json.dumps({'name':'abhi'})
r = requests.post(url,data)

print r.text

### Build ML API using flask

ml_api_script_file = "C:/Users/ABHISHEK/titanic/src/models/ml_api.py"

%%writefile $ml_api_script_file
from flask import Flask,request
import pandas as pd
import numpy as np
import pickle
import json
import os

app = Flask(__name__)

# Load Model and Scaler files
model_path = "C:/Users/ABHISHEK/titanic/models"
model_filepath = os.path.join(model_path,'lr_model.pkl')
scaler_filepath = os.path.join(model_path,'lr_scaler.pkl')

scaler = pickle.load(open(scaler_filepath))
model = pickle.load(open(model_filepath))

#columns

columns = [ u'Age', u'Fare', u'FamilySize', u'IsMother', u'IsMale',
       u'Deck_A', u'Deck_B', u'Deck_C', u'Deck_D', u'Deck_E', u'Deck_F',
       u'Deck_G', u'Deck_Z', u'Pclass_1', u'Pclass_2', u'Pclass_3',
       u'Title_Lady', u'Title_Master', u'Title_Miss', u'Title_Mr',
       u'Title_Mrs', u'Title_Officer', u'Title_Sir', u'Fare_Bin_very_low',
       u'Fare_Bin_low', u'Fare_Bin_high', u'Fare_Bin_very_high', u'Embarked_C',
       u'Embarked_Q', u'Embarked_S', u'AgeState_Adult', u'AgeState_Child']

@app.route('/api',methods=['POST'])
def make_prediction():
    # read json object and convert to json string
    data = json.dumps(requesr.get_json(force=True))
    # create pandas data frame using json string
    df = pd.read_json(data)
    # extract passenger ids
    passenger_ids = df['PassengerId'].ravel()
    # actual survived values
    actuals = df['Survived'].ravel()
    # extract required columns based and convert to matrix
    X = df[columns].as_matrix().astype('float')
    # transform the input
    X_scaled = scaler.transform(X)
    # make predictions
    predictions = model.predict(X_scaled)
    # create response dataframe
    df_response = pd.DataFrame({'PassengerId':passenger_ids,'Predicted':predictions,'Actual':actuals})
    # return json
    return df_response.to_json()

def run_app():
    app.run(port=10001,debug=True)
run_app()
